Refactoring Sat6 Activation Key
